genesisCoinbaseTxHex created automatically for new coins

+++ compile.sh
@@ -14,30 +14,16 @@ set -o errexit
 OPTIND=1         # Reset in case getopts has been used previously in the shell.
 
 
-while getopts "h?f:c:" opt; do
+while getopts "c:z" opt; do
     case "$opt" in
-    f)  CONFIG_FILE=${OPTARG}
-        ;;
     c)  COMPILE_ARGS=${OPTARG}
         ;;
+    z)  archive=1
     esac
 done
+archive=${archive:-0}
 
 shift $((OPTIND-1))
-
-if [ -z "${CONFIG_FILE}" ]; then
-    exit 1
-fi
-echo "compile:"
-echo ${CONFIG_FILE}
-echo ${COMPILE_ARGS}
-
-
-BUILD_PATH="${WORK_FOLDERS_PATH}/builds"
-MAC_BUILD_NAME="${__tick_data_core_CRYPTONOTE_NAME}-mac"
-LINUX_BUILD_NAME="${__tick_data_core_CRYPTONOTE_NAME}-linux"
-WINDOWS_BUILD_NAME="${__tick_data_core_CRYPTONOTE_NAME}-windows"
-
 
 cd ${NEW_COIN_PATH}
 rm -rf build; mkdir -p build/release; cd build/release
@@ -56,32 +42,39 @@ if [[ $? == "0" ]]; then
 fi
 
 # Move and zip binaries
-case "$OSTYPE" in
-  msys*) 	rm -f ${BUILD_PATH}/${WINDOWS_BUILD_NAME}.zip
-	rm -rf ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
-	mkdir -p ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
-	cp ${NEW_COIN_PATH}/build/release/src/${__tick_data_core_daemon_name} ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
-	cp ${NEW_COIN_PATH}/build/release/src/simplewallet ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
-	rm -rf "${NEW_COIN_PATH}/build"
-	cd ${BUILD_PATH}
-	zip -r ${WINDOWS_BUILD_NAME}.zip ${WINDOWS_BUILD_NAME}/
-	;;
-  darwin*)  	rm -f ${BUILD_PATH}/${MAC_BUILD_NAME}.zip
-	rm -rf ${BUILD_PATH}/${MAC_BUILD_NAME}
-	mkdir -p ${BUILD_PATH}/${MAC_BUILD_NAME}
-	cp ${NEW_COIN_PATH}/build/release/src/${__tick_data_core_daemon_name} ${BUILD_PATH}/${MAC_BUILD_NAME}
-	cp ${NEW_COIN_PATH}/build/release/src/simplewallet ${BUILD_PATH}/${MAC_BUILD_NAME}
-	rm -rf "${NEW_COIN_PATH}/build"
-	cd ${BUILD_PATH}
-	zip -r ${MAC_BUILD_NAME}.zip ${MAC_BUILD_NAME}/
-	;;
-  *)	rm -r ${BUILD_PATH}/${LINUX_BUILD_NAME}.tar.gz
-	rm -rf ${BUILD_PATH}/${LINUX_BUILD_NAME}
-	mkdir -p ${BUILD_PATH}/${LINUX_BUILD_NAME}
-	cp ${NEW_COIN_PATH}/build/release/src/${__tick_data_core_daemon_name} ${BUILD_PATH}/${LINUX_BUILD_NAME}
-	cp ${NEW_COIN_PATH}/build/release/src/simplewallet ${BUILD_PATH}/${LINUX_BUILD_NAME}
-	rm -rf "${NEW_COIN_PATH}/build"
-	cd ${BUILD_PATH}
-	tar -zcvf ${LINUX_BUILD_NAME}.tar.gz ${LINUX_BUILD_NAME}
-	;;
-esac
+if [[ $archive == "1" ]]; then
+	BUILD_PATH="${WORK_FOLDERS_PATH}/builds"
+	MAC_BUILD_NAME="${__tick_data_core_CRYPTONOTE_NAME}-mac"
+	LINUX_BUILD_NAME="${__tick_data_core_CRYPTONOTE_NAME}-linux"
+	WINDOWS_BUILD_NAME="${__tick_data_core_CRYPTONOTE_NAME}-windows"
+
+	case "$OSTYPE" in
+	  msys*) 	rm -f ${BUILD_PATH}/${WINDOWS_BUILD_NAME}.zip
+		rm -rf ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
+		mkdir -p ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
+		cp ${NEW_COIN_PATH}/build/release/src/${__tick_data_core_daemon_name} ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
+		cp ${NEW_COIN_PATH}/build/release/src/simplewallet ${BUILD_PATH}/${WINDOWS_BUILD_NAME}
+		rm -rf "${NEW_COIN_PATH}/build"
+		cd ${BUILD_PATH}
+		zip -r ${WINDOWS_BUILD_NAME}.zip ${WINDOWS_BUILD_NAME}/
+		;;
+	  darwin*)  	rm -f ${BUILD_PATH}/${MAC_BUILD_NAME}.zip
+		rm -rf ${BUILD_PATH}/${MAC_BUILD_NAME}
+		mkdir -p ${BUILD_PATH}/${MAC_BUILD_NAME}
+		cp ${NEW_COIN_PATH}/build/release/src/${__tick_data_core_daemon_name} ${BUILD_PATH}/${MAC_BUILD_NAME}
+		cp ${NEW_COIN_PATH}/build/release/src/simplewallet ${BUILD_PATH}/${MAC_BUILD_NAME}
+		rm -rf "${NEW_COIN_PATH}/build"
+		cd ${BUILD_PATH}
+		zip -r ${MAC_BUILD_NAME}.zip ${MAC_BUILD_NAME}/
+		;;
+	  *)	rm -r ${BUILD_PATH}/${LINUX_BUILD_NAME}.tar.gz
+		rm -rf ${BUILD_PATH}/${LINUX_BUILD_NAME}
+		mkdir -p ${BUILD_PATH}/${LINUX_BUILD_NAME}
+		cp ${NEW_COIN_PATH}/build/release/src/${__tick_data_core_daemon_name} ${BUILD_PATH}/${LINUX_BUILD_NAME}
+		cp ${NEW_COIN_PATH}/build/release/src/simplewallet ${BUILD_PATH}/${LINUX_BUILD_NAME}
+		rm -rf "${NEW_COIN_PATH}/build"
+		cd ${BUILD_PATH}
+		tar -zcvf ${LINUX_BUILD_NAME}.tar.gz ${LINUX_BUILD_NAME}
+		;;
+	esac
+fi
